+++ ReleaseNotes.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## 2.4.2
+* update whois_active property to return corrected value
+* updates for task filter
+* added malware and password parameters for document uploads
+
+## 2.4.1
+* Added task filter code
+
 ## 2.4
 * Added Tasks functionality
 * Update Victims functionlity